requirements.txt
djangorestframework
Django
mysqlclient
PyJWT
gunicorn